Tratando e lançando exceções.

+++ webservice/src/br/com/geekcode/webservices/model/dao/ProdutoDAO.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import javax.persistence.EntityManager;
 
+import br.com.geekcode.webservices.exceptions.DAOException;
+import br.com.geekcode.webservices.exceptions.ErrorCode;
 import br.com.geekcode.webservices.model.domain.Produto;
 
 public class ProdutoDAO {
@@ -14,6 +16,10 @@ public class ProdutoDAO {
 
 		try {
 			produtos = em.createQuery("select p from Produto p", Produto.class).getResultList();
+			
+		} catch (RuntimeException e) {
+			throw new DAOException("Erro ao recuperar todos os produtos do banco: " + e.getMessage(), ErrorCode.SERVER_ERROR.getCode());
+			
 		} finally {
 			em.close();
 		}
@@ -25,26 +31,77 @@ public class ProdutoDAO {
 		EntityManager em = JPAUtil.getEntityManager();
 		Produto produto = null;
 		
+		if (id <= 0) {
+			throw new DAOException("O id precisa ser maior que 0.", ErrorCode.BAD_REQUEST.getCode());
+		}
+		
 		try {
 			produto = em.find(Produto.class, id);
+			
+		} catch (RuntimeException e) {
+			throw new DAOException("Erro ao buscar produto por id no banco de dados: " + e.getMessage(), ErrorCode.SERVER_ERROR.getCode());
+			
 		} finally {
 			em.close();
 		}
 		
+		if (produto == null) {
+			throw new DAOException("Produto de id " + id + " não existe.", ErrorCode.NOT_FOUND.getCode());
+		}
+		
 		return produto;
+	}
+	
+	public Produto save(Produto produto) {
+		EntityManager em = JPAUtil.getEntityManager();
+		
+		if (!produtoIsValid(produto)) {
+			throw new DAOException("Produto com dados incompletos.", ErrorCode.BAD_REQUEST.getCode());
+		}
+		
+		try {
+			em.getTransaction().begin();
+			em.persist(produto);
+			em.getTransaction().commit();
+			
+		} catch (RuntimeException e) {
+			em.getTransaction().rollback();
+			throw new DAOException("Erro ao salvar produo no banco de dados: " + e.getMessage(), ErrorCode.SERVER_ERROR.getCode());
+			
+		} finally {
+			em.close();
+		}
+		
+		return produto;		
 	}
 	
 	public Produto update(Produto produto) {
 		EntityManager em = JPAUtil.getEntityManager();
 		Produto produtoManaged = null;
 		
+		if (produto.getId() <= 0) {
+			throw new DAOException("O id precisa ser maior que 0.", ErrorCode.BAD_REQUEST.getCode());
+		}
+		
+		if (!produtoIsValid(produto)) {
+			throw new DAOException("Produto com dados incompletos.", ErrorCode.BAD_REQUEST.getCode());
+		}
+		
 		try {
 			em.getTransaction().begin();
 			produtoManaged = em.find(Produto.class, produto.getId());
 			produtoManaged.setNome(produto.getNome());
 			produtoManaged.setQuantidade(produto.getQuantidade());
-			//em.merge(produtoManaged);
 			em.getTransaction().commit();
+			
+		} catch (NullPointerException e) {
+			em.getTransaction().rollback();
+			throw new DAOException("Produto informado para atualização não existe: " + e.getMessage(), ErrorCode.NOT_FOUND.getCode());
+			
+		} catch (RuntimeException e) {
+			em.getTransaction().rollback();
+			throw new DAOException("Erro ao atualizar produto no banco de dados: " + e.getMessage(), ErrorCode.SERVER_ERROR.getCode());
+			
 		} finally {
 			em.close();
 		}
@@ -52,15 +109,29 @@ public class ProdutoDAO {
 		return produtoManaged;
 	}
 	
+
 	public Produto delete(Long id) {
 		EntityManager em = JPAUtil.getEntityManager();
-		Produto produto = null
+		Produto produto = null;
+		
+		if (id <= 0) {
+			throw new DAOException("O id precisa ser maior que 0.", ErrorCode.BAD_REQUEST.getCode());
+		}
 				
 		try {
 			em.getTransaction().begin();
 			produto = em.find(Produto.class, id);
 			em.remove(produto);
 			em.getTransaction().commit();
+			
+		} catch (IllegalArgumentException e) {
+			em.getTransaction().rollback();
+			throw new DAOException("Produto informado para a remoção não existe: " + e.getMessage(), ErrorCode.NOT_FOUND.getCode());
+			
+		} catch (RuntimeException e) {
+			em.getTransaction().rollback();
+			throw new DAOException("Erro ao remover produto do banco de dados." + e.getMessage(), ErrorCode.SERVER_ERROR.getCode());
+			
 		} finally {
 			em.close();
 		}
@@ -68,4 +139,17 @@ public class ProdutoDAO {
 		return produto;
 	}
 
+	private boolean produtoIsValid(Produto produto) {
+		
+		try {
+			if ((produto.getNome().isEmpty())  || (produto.getQuantidade() < 0)) {
+				return false;
+			}
+			
+		} catch (NullPointerException e) {
+			throw new DAOException("Produto com dados incompletos. " + e.getMessage(), ErrorCode.BAD_REQUEST.getCode());
+		}		
+
+		return true;
+	}
 }
